+++ R/tidysmd-package.R
@@ -3,6 +3,8 @@
 
 ## usethis namespace: start
 #' @import rlang
+#' @importFrom stats model.matrix
 #' @importFrom stats weights
+#' @importFrom utils packageVersion
 ## usethis namespace: end
 NULL