+++ R/utils.R
@@ -4,5 +4,15 @@ globalVariables(c(
   "smd",
   "estimate",
   "smd_results",
-  "term"
+  "term",
+  "method",
+  "(Intercept)"
 ))
+
+warn <- function(...) {
+  rlang::warn(..., use_cli_format = TRUE, class = "tidysmd_warning")
+}
+
+abort <- function(...) {
+  rlang::abort(..., use_cli_format = TRUE, class = "tidysmd_error")
+}