+++ NEWS.md
@@ -1,5 +1,12 @@
+# tidysmd 0.2.0
+* Added a `make_dummy_vars` argument to `tidy_smd()` to transform categorical variables to dummy variables prior to calculating the SMD (#7)
+* Added `geom_love()` (#5)
+* Added `bind_matches()` and improved support for matched datasets (#10). Some arguments and output were changed to support this generalization and may result in breaking changes.
+
 # tidysmd 0.1.1
 
-# tidysmd 0.1.0.9000
+* Fixed imports issue related to tidyselect
+
+# tidysmd 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.